pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-cs040-chisatokamakura"
version = "0.1.3"
description = "LLM-based chat/document summarization project for CSC040"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq",
    "python-dotenv",
    "GitPython",
]

[project.scripts]
chat = "chat:main"

[tool.flake8]
exclude = ["venv", ".venv", "__pycache__", "build", "test_projects"]

[tool.setuptools]
py-modules = ["chat"]

[tool.setuptools.packages.find]
include = ["tools*"]
exclude = ["test_projects*", "build*", "dist*", "images*"]

[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=test_utf16.txt"